// src/domain/types/doctorSpecialty.type.ts
export enum DoctorSpecialty {
  CARDIOLOGY = 'CARDIOLOGY',
  DERMATOLOGY = 'DERMATOLOGY',
  PEDIATRICS = 'PEDIATRICS',
  NEUROLOGY = 'NEUROLOGY',
  PSYCHIATRY = 'PSYCHIATRY',
  RADIOLOGY = 'RADIOLOGY',
  SURGERY = 'SURGERY',
  INTERNAL_MEDICINE = 'INTERNAL_MEDICINE',
  EMERGENCY_MEDICINE = 'EMERGENCY_MEDICINE',
  ANESTHESIOLOGY = 'ANESTHESIOLOGY'
}